refactor: extract the DNS/process subsystem into DNSResolver (US-013)
Continue the god-class split: move the pure DNS resolvers (resolveIPsParallel,
resolveWithDNSParallel/DoTParallel/DoHParallel, resolveWithSystemResolver,
isValidIPStatic) and the process runners (runProcessParallel, runProcessSyncSafe)
out of RouteManager into a top-level `enum DNSResolver`. These were already
nonisolated/static (side-effect-free); this just groups them into a collaborator
so RouteManager shrinks and the DNS layer is a standalone unit.

Pure code motion (bodies verbatim) + call-site requalification (Self.X -> DNSResolver.X);
the shared process queue and two cross-type entry points widened from private to
module-internal. Behavior-preserving. Build clean, 786 tests / 0 failures.
RouteManager.swift 3955 -> 3671 lines.
